Started writing a test file for Eval_Deriv
I want to test that this function actually works before moving onto
other stuff. To do this, I wrote a simple function (a polynomial of
x, y, and t), with known derivatives. I'll use this to test that the
code works in 1 and 2 spatial dimensions.

I've written part of the 2d case... I need to add in actual unit tests,
and then I'll write the 1d case.