Finished tests for Evaluate Derivatives
So far, so good. It looks like the function works! I had to fix several
bugs, but the core code seems to be good.

Up next, I want to write functions for book-keeping in the library.
In particular, I want a function that, given a number of x and y
derivatives, tells you which "index value" that spatial partial
derivative corresponds to in a multi-index.

I then want a function which maps a multi-index to the corresponding
component of Xi (equivalently, column of the library)

I then want functions that invert both operations.

These functions should make it easy to map construct polynomial terms
(by mapping the sub-indices in the multi-index back to their
corresponding derivatives), construct the approximation to du/dt (by
multiplying a polynomial term by its corresponding component of Xi
and then adding the result to the approximation of dU/dt), and mapping
a learned PDE (in Xi) back to a human-readable form (by mapping the
non-zero components to their polynomial forms, then representing each
polynomial form as a product of derivatives).

Once I have these mapping functions, constructing the actual library
should be pretty straight forward. Expect code to test them before
I write the library, however.