Wrote some library functions, tests
I wrote functions to map a given number of x, y derivatives to an
index value and vice versa. These functions basically tell me which
spatial partial derivative a given multi-index sub-index value
corresponds to. This will allow me to determine which library
term (in human-readable form) a multi-index corresponds to.

I also wrote a functor to map a multi-index to a column number. This
proved a little tricky, and I had to use a big look-up table. I'm
going to have to do something similar to map a column number back to a
multi-index.... and I'm thinking of replacing my index value to
number of x, y derivatives array thing into a functor as well.

Expect all of that, along with tests, in the next commit or two.

After that, implementing the code should be (relatively) smooth
sailing.