Made Index_to_xy_Derivatives into a functor
I also updated the tests to reflect this. I also changed the
xy_Derivatives_To_Index test to included randomness, and to use the
fact that I don't care how xy_Derivatives_To_Index maps (i,j) pairs to
elements of N, so long as each i,j with i + j <= N get mapped (in an
injective way) to {0, 1, 2,.... (N + 1)(N + 2)/2 - 1}. The check now
reflects this assumption.

I believe my book-keeping functions are now complete. The next step is
to construct the actual library, as well as a function that maps each
multi-index to a human-readable from of the encoded term.