Wrote Coll, Lp Loss functions.
I believe the work now. The Collocation Loss should work for 1 or 2
spatial dimensions.

I also used this opportunity to modify a few aspects of the mappings.
For technical reasons, it made more sense to have a dedicated function
for mapping Indices to x derivatives (even though this is literally
just the identity). So, I wrote such a function. I also modified
Index_To_xy_Derivatives such that it always returns a 2 element numpy
array, not a 2 by 1. I also modified Col_Number_To_Multi_Index such
that the returned multi-index is an N element array (where N is the
number of sub-indices) rather than a N by 1. These changes are subtle,
but made the Coll_Loss function a little cleaner.

I also modified Evaluate_Derivatives such that Dt_U is an N element
array (where N is the number of coordinates), rather than a N by 1. This
change again makes the Coll_Loss function a little cleaner.

I made most of the above changes so that the shape of the returned
arrays are more "natural" (as defined by what I felt was natural).

None of the new code is tested. I want to write a test for the loss
functions using my Polynomial classes that I used to test
Evaluate_Derivatives. Once those tests are done and running, I believe
the code will be basically ready to go. I need to write a main function,
and a function to obtain data from file... but otherwise I think
everything is here.