Added improved Lp loss + tests.
I realized I was being dumb yesterday. We know that if x != 0, then,
      lim_{p -> 0} |x|^p = 1
While if x = 0, then
      lim_{p -> 0} |x|^p = 0
Thus,
      |Xi|_0 = \lim_{p -> 0} |Xi[0]|^p + ... + |Xi[N-1]|^p
My mistake before was raising the result to the power of 1/p... that's
dumb. I decided to keep the thing I came up with yesterday as well. I
think the new thing works better, but want to test both. I also
like the idea of gradually decreasing p as the epochs progress to
"sweep" all the un-important terms in the library to zero.

In the next commit, I plan to split the test file into a bunch of sub
files... it's getting out of hand in one file.