Added new Xi file + thresholding
I added a new file to deal with Xi specific functions... namely
converting Xi into a printable PDE. I also wrote a function to do
thresholding based on the product |Xi[k]|*(2 norm of kth lib term).
This approach seems to make sense, but fails in practice. I observed
that the higher the total number of derivatives in the lib term, the
larger its 2 norm. As such, this approach to thresholding fails
miserably. That shouldn't be too surprising. Our algorithm attempts to
minimize the size of the components of Xi (via simulated Lp regression).
There is no least squares in sight... The 2 norm trick works when we
want to minimize how much we increase the least-squares residual. That
doesn't appear to work in this case. As a result, I choose to threshold
based purely on the magnitude of the components of Xi.

I also updated Print_PDE (now in the Xi.py file) such that it prints
U instead of D_x^0 U.

I still want to update the Lp loss test and add a plotter. Expect those
changes soonish.