Fixed bugs for 2d datasets + added descriptions
I found several bugs in the Data_Loader for data sets with 2 spatial
variables. I should have expected this given that this was one of the
few pieces of code that I didn't unit test. This really shows how
powerful unit testing is.

I also added code to "Frankenstein" a data set with one spatial variable
to one with 2 spatial variables. This works by generating a new data
set such that the (t, x, y) coordinate of the 2d one is the (t, x)
coordinate of the 1d one. In other words, it copies the 1d data set for
every y coordinates. The code is slow, inefficient, and ugly. However,
it's sorta supposed to be given what it does. I added this feature
(which is currently commented out) so that I can test the 2d code using
just 1d data sets... further, the solution PDE should "learn" to ignore
the second spatial variable (as it should!).

I also added descriptions to "Print_PDE" and "Coll_Loss". In my next
commit, I will update the Lp loss test code.