Updated Test_Lp_Loss
I didn't actually have to change any code... I just modified a few
comments. With this commit, I believe my code is "done" in the sense
that I've added all features that I originally planned to add. I will
eventually add a plotter for plotting solutions to 1d data sets, though
that's not urgent. The important thing is that the code works, and
appears to work well.

I also plan to eventually play around with L0_Approx_Loss and modifying
the value of s... However, expect that to happen later. In the mean
time, I'm ver happy with the state of my code.