Switched Xi to float32
I am not sure what's going wrong. If I save the state when using the
the LBFGS optimizer, and subsequently load from save, I get the
following error:

Traceback (most recent call last):
  File "./main.py", line 268, in <module>
    main();
  File "./main.py", line 179, in main
    Device                      = Settings.Device);
  File "/Users/robertstephany/Documents/Grad school/Classes/Math
        Modeling/Project/DeepMoD_PlusPlus/Code/Test_Train.py",
        line 115, in Training
    Optimizer.step(Closure);
  File "/Users/robertstephany/Library/Python/3.7/lib/python/
        site-packages/torch/optim/optimizer.py",
        line 89, in wrapper
    return func(*args, **kwargs)
  File "/Users/robertstephany/Library/Python/3.7/lib/python/
        site-packages/torch/autograd/grad_mode.py",
        line 27, in decorate_context
    return func(*args, **kwargs)
  File "/Users/robertstephany/Library/Python/3.7/lib/python/
        site-packages/torch/optim/lbfgs.py", line 353, in step
    ys = y.dot(s)  # y*s
RuntimeError: dot : expected both vectors to have same dtype, but found
Double and Float.

I have no clue what's causing this. I tried recreating the problem in a
simple bit of code, but could not. I plan to gradually add
detail to my simple code until the error reappears.

In the mean time, the error appears to go away if I make Xi a float
tensor... so that's what I'm doing. I want to understand what's going
wrong, however, so I plan to address this in the future. This is
merely a patch for now.