Added SIREN support
I added support for SIRENS (basically NN's with the sin activation function). I did this to try and learn the KS equation. It appears to have failed. I also had to change the initialization scheme, because the one proposed in the paper causes U's derivatives to explode. I, therefore, initialize the first weight matrix to be all zero. This seems to
work. The network does alright, but still can't learn the KS equation. Oh well. I kept the changes because they may work well for other data sets.