Changed the way that PDE_LEARN takes in data
I completely rewrote the way that PDE_LEARN takes in data. Before, the function Data_Loader read in a .mat file, and then added sparsity and noise to the dataset within. This worked well, but created a big problem: PDE_LEARN could only train on data sets that were generated in MATLAB and used a very specific naming system for the saved data. This basically means that PDE_LEARN can not easily train on data from anything other than my matlab scripts. This is a big problem. The last two commits eliminate this limitation.

The big change is that PDE_LEARN no longer adds sparsity and noise. Instead, it reads in a .npz file that contains testing/training inputs/targets, and information about the problem domain. Sparsity and noise is built into the data set. All data sets use the same universal format. This change may seem simple, but its significant. Going forward, PDE_LEARN does not make data sets, it only reads them. It is up to you, the user, to supply code that converts your data into the universal format. I wrote code to do this for the data sets produced by my MATLAB scripts. In particular, the file From_MATLAB reads in a .mat file, processes it (doing much of the work that the old Data_Loader function did), and then passes the results to Create_DataSet which serializes it using the format that PDE_LEARN expects.

This approach should make it much easier to use PDE_LEARN with other data sets. It also means that PDE_LEARN has fewer settings... a lot fewer (no noise level, # of test/train points, or # of spatial dimensions... The DataSet contains all this information).

I have tested the new approach on a Burgers equation experiment, and everything appears to be running smoothly.