Added targeted collocation points
It was relatively straight-forward. I choose to use the following approach: at each epoch, generate a new set of collocation points by sampling a uniform distribution over the problem domain. I then evaluate the collocation loss at these points, as well as the current targeted points (we make the targeted set empty to start). Look at the mean and SD of the absolute PDE residual (abs value of PDE residual) at the random points. Determine which collocation points (targeted or random) whose absolute residual is more than 2 SDs from the mean. These points become the new targeted points.

I like this approach for a few reasons: one, there are no tunable parameters; the cutoff for what residual is too larger (more than 2sd above the mean) changes as the network trains. Second, it focuses the targeted points in the "extreme" parts of the domain; places where the PDE residual is gigantic. This helps ensure that the collocation loss never gets too big in any one region of the domain.

To make this work, I had to modify the Coll_Loss function. It now returns two arguments. The first is the collocation loss, the second is the PDE Residual at the collocation coordinates. Test_Train uses the first argument for training, but returns the second argument (so that the main function can determine the new targeted points). I don't claim that 2 SDs from the mean is ideal... but it seems to work.

In my next commit, I plan to greatly enhance the plotting capabilities of PDE-LEARN (by plotting the PDE Residual)