Fixed buggs, got code running with new save method
I think it works. I need to do some more testing to make sure.

The big idea of the new system is that the code saves everything it needs to restart where it left off. You don't need to remember the network architecture, or the library terms... all of that is saved with the network parameters. If you load everything from save, you can ignore most of the settings. This should make the code way way way easier to use.