Added L2 Loss
I added L2 loss (for network parameters). I also got rid of the L0_Approx_Loss... since I do not plan to use it.

I modified the rest of the code to work with the L2 loss + added it to the Settings.

In my next commit, I plan to add a function that adds support for plotting loss history