Another round of edits on the README
I did a full read-through of the README. I found several minor typos, rewored several things, and corrected one major typo. The description of `Testing` and `Training` in the old README was completely wrong. I fixed it.

This is likely to be the last major update of the README. I will probably tweak it going forward, but think it is sufficiently refined.